Updated how the form behaves and displays
added JS to make bkgrnd transparent after user enters data. And also cleared all the input features when the Save button is pressed to simulate an update to the server by displaying only the data. When done, this page can be used to add/update/delete a recipe and also to display only recipe data, that way it can have multi-purpose by dynamically altering the page as needed per function type (CRUD).